within app, moved all the bip classes into bip package
