+++ app/src/main/java/gov/va/vro/bip/service/BipApiService.java
@@ -1,7 +1,9 @@
-package gov.va.vro.service.provider.bip.service;
+package gov.va.vro.bip.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import gov.va.vro.bip.BipApiProps;
+import gov.va.vro.bip.BipException;
 import gov.va.vro.model.bip.BipClaim;
 import gov.va.vro.model.bip.BipClaimResp;
 import gov.va.vro.model.bip.BipContentionResp;
@@ -9,8 +11,6 @@ import gov.va.vro.model.bip.BipUpdateClaimResp;
 import gov.va.vro.model.bip.ClaimContention;
 import gov.va.vro.model.bip.ClaimStatus;
 import gov.va.vro.model.bip.UpdateContentionReq;
-import gov.va.vro.service.provider.BipApiProps;
-import gov.va.vro.service.provider.bip.BipException;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Header;
 import io.jsonwebtoken.Jwts;
